pyproject.toml
[project]
name = "confluence-export"
version = "0.1.0"
description = "LLM-ready Confluence page tree exporter"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "markdownify>=0.11",
    "pyyaml>=6.0",
]

[project.scripts]
confluence-export = "confluence_export.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/confluence_export"]

[tool.pytest.ini_options]
testpaths = ["tests"]
